Refactor operators
