+++ brainpy/_src/math/event/__init__.py
@@ -1,5 +1,4 @@
 
 from ._info_collection import *
 from ._csr_matvec import *
-from ._csr_matvec_taichi import *
