+++ brainpy/_src/math/event/tests/test_event_csrmv.py
@@ -16,6 +16,8 @@ is_manual_test = False
 if platform.system() == 'Windows' and not is_manual_test:
   pytest.skip('brainpy.math package may need manual tests.', allow_module_level=True)
 
+brainpylib_csr_matvec = partial(bm.event.csrmv, method='brainpylib')
+taichi_csr_matvec = partial(bm.event.csrmv, method='taichi')
 
 def sum_op(op):
   def func(*args, **kwargs):
@@ -56,18 +58,18 @@ class Test_event_csr_matvec(parameterized.TestCase):
     events = rng.random(shape[0] if transpose else shape[1]) < 0.1
     heter_data = bm.ones(indices.shape) * homo_data
 
-    r1 = bm.event.csrmv(homo_data, indices, indptr, events, shape=shape, transpose=transpose)
-    r2 = bm.event.csrmv(heter_data, indices, indptr, events, shape=shape, transpose=transpose)
+    r1 = brainpylib_csr_matvec(homo_data, indices, indptr, events, shape=shape, transpose=transpose)
+    r2 = brainpylib_csr_matvec(heter_data, indices, indptr, events, shape=shape, transpose=transpose)
     self.assertTrue(bm.allclose(r1, r2))
 
-    r3 = bm.event.csrmv(homo_data, indices, indptr, events.astype(float), shape=shape, transpose=transpose)
+    r3 = brainpylib_csr_matvec(homo_data, indices, indptr, events.astype(float), shape=shape, transpose=transpose)
     self.assertTrue(bm.allclose(r1, r3))
 
     dense = bm.sparse.csr_to_dense(heter_data, indices, indptr, shape=shape)
     r4 = (events @ dense) if transpose else (dense @ events)
     self.assertTrue(bm.allclose(r1, r4))
 
-    r5 = bm.event.csrmv(heter_data, indices, indptr, events.astype(float), shape=shape, transpose=transpose)
+    r5 = brainpylib_csr_matvec(heter_data, indices, indptr, events.astype(float), shape=shape, transpose=transpose)
     self.assertTrue(bm.allclose(r1, r5))
 
     bm.clear_buffer_memory()
@@ -97,7 +99,7 @@ class Test_event_csr_matvec(parameterized.TestCase):
 
     # vmap 'data'
     events = bm.as_jax(rng.random(shape[0] if transpose else shape[1])) < 0.1
-    f1 = jax.vmap(partial(bm.event.csrmv, indices=indices, indptr=indptr, events=events,
+    f1 = jax.vmap(partial(brainpylib_csr_matvec, indices=indices, indptr=indptr, events=events,
                           shape=shape, transpose=transpose))
     f2 = jax.vmap(
       partial(partial(bm.sparse.csrmv, method='cusparse'), indices=indices, indptr=indptr, vector=events.astype(float),
@@ -106,7 +108,7 @@ class Test_event_csr_matvec(parameterized.TestCase):
     self.assertTrue(bm.allclose(f1(vmap_data), f2(vmap_data)))
 
     # vmap 'events'
-    f3 = jax.vmap(partial(bm.event.csrmv, homo_data, indices, indptr,
+    f3 = jax.vmap(partial(brainpylib_csr_matvec, homo_data, indices, indptr,
                           shape=shape, transpose=transpose))
     f4 = jax.vmap(partial(partial(bm.sparse.csrmv, method='cusparse'), homo_data, indices, indptr,
                           shape=shape, transpose=transpose))
@@ -114,7 +116,7 @@ class Test_event_csr_matvec(parameterized.TestCase):
     self.assertTrue(bm.allclose(f3(vmap_data), f4(vmap_data.astype(float))))
 
     # vmap 'data' and 'events'
-    f5 = jax.vmap(lambda dd, ee: bm.event.csrmv(dd, indices, indptr, ee, shape=shape, transpose=transpose))
+    f5 = jax.vmap(lambda dd, ee: brainpylib_csr_matvec(dd, indices, indptr, ee, shape=shape, transpose=transpose))
     f6 = jax.vmap(lambda dd, ee: bm.sparse.csrmv(dd, indices, indptr, ee, shape=shape, transpose=transpose,
                                                  method='cusparse'))
     vmap_data1 = bm.as_jax([homo_data] * 10)
@@ -152,7 +154,7 @@ class Test_event_csr_matvec(parameterized.TestCase):
     dense_conn = bm.sparse.csr_to_dense(bm.ones(indices.shape).value, indices, indptr, shape=shape)
 
     # grad 'data'
-    r1 = jax.grad(sum_op(bm.event.csrmv))(
+    r1 = jax.grad(sum_op(brainpylib_csr_matvec))(
       homo_data, indices, indptr, events, shape=shape, transpose=transpose)
     r2 = jax.grad(sum_op(partial(bm.sparse.csrmv, method='cusparse')))(
       homo_data, indices, indptr, events.astype(float), shape=shape, transpose=transpose)
@@ -162,7 +164,7 @@ class Test_event_csr_matvec(parameterized.TestCase):
     self.assertTrue(bm.allclose(r1, r3))
 
     # grad 'events'
-    r4 = jax.grad(sum_op(bm.event.csrmv), argnums=3)(
+    r4 = jax.grad(sum_op(brainpylib_csr_matvec), argnums=3)(
       homo_data, indices, indptr, events.astype(float), shape=shape, transpose=transpose)
     r5 = jax.grad(sum_op(partial(bm.sparse.csrmv, method='cusparse')), argnums=3)(
       homo_data, indices, indptr, events.astype(float), shape=shape, transpose=transpose)
@@ -198,7 +200,7 @@ class Test_event_csr_matvec(parameterized.TestCase):
     events = bm.as_jax(rng.random(shape[0] if transpose else shape[1])) < 0.1
     heter_data = bm.as_jax(rng.random(indices.shape))
 
-    r1 = bm.event.csrmv(heter_data, indices, indptr, events,
+    r1 = brainpylib_csr_matvec(heter_data, indices, indptr, events,
                         shape=shape, transpose=transpose)
     r2 = partial(bm.sparse.csrmv, method='cusparse')(heter_data, indices, indptr, events.astype(float),
                                                      shape=shape, transpose=transpose)
@@ -208,7 +210,7 @@ class Test_event_csr_matvec(parameterized.TestCase):
     r3 = (events @ dense) if transpose else (dense @ events)
     self.assertTrue(bm.allclose(r1, r3))
 
-    r4 = bm.event.csrmv(heter_data, indices, indptr, events.astype(float),
+    r4 = brainpylib_csr_matvec(heter_data, indices, indptr, events.astype(float),
                         shape=shape, transpose=transpose)
     self.assertTrue(bm.allclose(r1, r4))
 
@@ -239,7 +241,7 @@ class Test_event_csr_matvec(parameterized.TestCase):
 
     # vmap 'data'
     events = bm.as_jax(rng.random(shape[0] if transpose else shape[1])) < 0.1
-    f1 = jax.vmap(partial(bm.event.csrmv, indices=indices, indptr=indptr, events=events,
+    f1 = jax.vmap(partial(brainpylib_csr_matvec, indices=indices, indptr=indptr, events=events,
                           shape=shape, transpose=transpose))
     f2 = jax.vmap(
       partial(partial(bm.sparse.csrmv, method='cusparse'), indices=indices, indptr=indptr, vector=events.astype(float),
@@ -249,7 +251,7 @@ class Test_event_csr_matvec(parameterized.TestCase):
 
     # vmap 'events'
     data = bm.as_jax(rng.random(indices.shape))
-    f3 = jax.vmap(partial(bm.event.csrmv, data, indices, indptr,
+    f3 = jax.vmap(partial(brainpylib_csr_matvec, data, indices, indptr,
                           shape=shape, transpose=transpose))
     f4 = jax.vmap(partial(partial(bm.sparse.csrmv, method='cusparse'), data, indices, indptr,
                           shape=shape, transpose=transpose))
@@ -257,7 +259,7 @@ class Test_event_csr_matvec(parameterized.TestCase):
     self.assertTrue(bm.allclose(f3(vmap_data), f4(vmap_data.astype(float))))
 
     # vmap 'data' and 'events'
-    f5 = jax.vmap(lambda dd, ee: bm.event.csrmv(dd, indices, indptr, ee,
+    f5 = jax.vmap(lambda dd, ee: brainpylib_csr_matvec(dd, indices, indptr, ee,
                                                 shape=shape, transpose=transpose))
     f6 = jax.vmap(lambda dd, ee: partial(bm.sparse.csrmv, method='cusparse')(dd, indices, indptr, ee,
                                                                              shape=shape, transpose=transpose))
@@ -295,7 +297,7 @@ class Test_event_csr_matvec(parameterized.TestCase):
 
     # grad 'data'
     data = bm.as_jax(rng.random(indices.shape))
-    r1 = jax.grad(sum_op(bm.event.csrmv))(
+    r1 = jax.grad(sum_op(brainpylib_csr_matvec))(
       data, indices, indptr, events, shape=shape, transpose=transpose)
     r2 = jax.grad(sum_op(partial(bm.sparse.csrmv, method='cusparse')))(
       data, indices, indptr, events.astype(float), shape=shape, transpose=transpose)
@@ -309,7 +311,7 @@ class Test_event_csr_matvec(parameterized.TestCase):
     self.assertTrue(bm.allclose(r1, r3))
 
     # grad 'events'
-    r4 = jax.grad(sum_op(bm.event.csrmv), argnums=3)(
+    r4 = jax.grad(sum_op(brainpylib_csr_matvec), argnums=3)(
       data, indices, indptr, events.astype(float), shape=shape, transpose=transpose)
     r5 = jax.grad(sum_op(partial(bm.sparse.csrmv, method='cusparse')), argnums=3)(
       data, indices, indptr, events.astype(float), shape=shape, transpose=transpose)
