+++ brainpy/_src/math/event/tests/test_event_csrmv_taichi.py
@@ -19,6 +19,8 @@ def sum_op(op):
 
   return func
 
+brainpylib_csr_matvec = partial(bm.event.csrmv, method='brainpylib')
+taichi_csr_matvec = partial(bm.event.csrmv, method='taichi')
 
 class Test_event_csr_matvec_taichi(parameterized.TestCase):
   def __init__(self, *args, platform='cpu', **kwargs):
@@ -42,8 +44,8 @@ class Test_event_csr_matvec_taichi(parameterized.TestCase):
     events = rng.random(shape[0] if transpose else shape[1]) < 0.1
     heter_data = bm.ones(indices.shape) * homo_data
 
-    r1 = bm.event.csrmv(homo_data, indices, indptr, events, shape=shape, transpose=transpose)
-    r2 = bm.event.csrmv_taichi(homo_data, indices, indptr, events, shape=shape, transpose=transpose)
+    r1 = brainpylib_csr_matvec(homo_data, indices, indptr, events, shape=shape, transpose=transpose)
+    r2 = taichi_csr_matvec(homo_data, indices, indptr, events, shape=shape, transpose=transpose)
 
     assert (bm.allclose(r1, r2))
 
@@ -65,24 +67,24 @@ class Test_event_csr_matvec_taichi(parameterized.TestCase):
 
     # vmap 'data'
     events = bm.as_jax(rng.random(shape[0] if transpose else shape[1])) < 0.1
-    f1 = jax.vmap(partial(bm.event.csrmv, indices=indices, indptr=indptr, events=events,
+    f1 = jax.vmap(partial(brainpylib_csr_matvec, indices=indices, indptr=indptr, events=events,
                           shape=shape, transpose=transpose))
-    f2 = jax.vmap(partial(bm.event.csrmv_taichi, indices=indices, indptr=indptr, events=events,
+    f2 = jax.vmap(partial(taichi_csr_matvec, indices=indices, indptr=indptr, events=events,
                           shape=shape, transpose=transpose))
     vmap_data = bm.as_jax([homo_data] * 10)
     self.assertTrue(bm.allclose(f1(vmap_data), f2(vmap_data)))
 
     # vmap 'events'
-    f3 = jax.vmap(partial(bm.event.csrmv, homo_data, indices, indptr,
+    f3 = jax.vmap(partial(brainpylib_csr_matvec, homo_data, indices, indptr,
                           shape=shape, transpose=transpose))
-    f4 = jax.vmap(partial(bm.event.csrmv_taichi, homo_data, indices, indptr,
+    f4 = jax.vmap(partial(taichi_csr_matvec, homo_data, indices, indptr,
                           shape=shape, transpose=transpose))
     vmap_data = bm.as_jax(rng.random((10, shape[0] if transpose else shape[1]))) < 0.1
     self.assertTrue(bm.allclose(f3(vmap_data), f4(vmap_data)))
 
     # vmap 'data' and 'events'
-    f5 = jax.vmap(lambda dd, ee: bm.event.csrmv(dd, indices, indptr, ee, shape=shape, transpose=transpose))
-    f6 = jax.vmap(lambda dd, ee: bm.event.csrmv_taichi(dd, indices, indptr, ee, shape=shape, transpose=transpose))
+    f5 = jax.vmap(lambda dd, ee: brainpylib_csr_matvec(dd, indices, indptr, ee, shape=shape, transpose=transpose))
+    f6 = jax.vmap(lambda dd, ee: taichi_csr_matvec(dd, indices, indptr, ee, shape=shape, transpose=transpose))
 
     vmap_data1 = bm.as_jax([homo_data] * 10)
     vmap_data2 = bm.as_jax(rng.random((10, shape[0] if transpose else shape[1]))) < 0.2
@@ -110,16 +112,16 @@ class Test_event_csr_matvec_taichi(parameterized.TestCase):
     dense_conn = bm.sparse.csr_to_dense(bm.ones(indices.shape).value, indices, indptr, shape=shape)
 
     # grad 'data'
-    r1 = jax.grad(sum_op(bm.event.csrmv))(
+    r1 = jax.grad(sum_op(brainpylib_csr_matvec))(
       homo_data, indices, indptr, events, shape=shape, transpose=transpose)
-    r2 = jax.grad(sum_op(bm.event.csrmv_taichi))(
+    r2 = jax.grad(sum_op(taichi_csr_matvec))(
       homo_data, indices, indptr, events, shape=shape, transpose=transpose)
     self.assertTrue(bm.allclose(r1, r2))
 
     # grad 'events'
-    r3 = jax.grad(sum_op(bm.event.csrmv), argnums=3)(
+    r3 = jax.grad(sum_op(brainpylib_csr_matvec), argnums=3)(
       homo_data, indices, indptr, events.astype(float), shape=shape, transpose=transpose)
-    r4 = jax.grad(sum_op(bm.event.csrmv_taichi), argnums=3)(
+    r4 = jax.grad(sum_op(taichi_csr_matvec), argnums=3)(
       homo_data, indices, indptr, events.astype(float), shape=shape, transpose=transpose)
     self.assertTrue(bm.allclose(r3, r4))
 
@@ -141,9 +143,9 @@ class Test_event_csr_matvec_taichi(parameterized.TestCase):
     events = bm.as_jax(rng.random(shape[0] if transpose else shape[1])) < 0.1
     heter_data = bm.as_jax(rng.random(indices.shape))
 
-    r1 = bm.event.csrmv(heter_data, indices, indptr, events,
+    r1 = brainpylib_csr_matvec(heter_data, indices, indptr, events,
                         shape=shape, transpose=transpose)
-    r2 = bm.event.csrmv_taichi(heter_data, indices, indptr, events,
+    r2 = taichi_csr_matvec(heter_data, indices, indptr, events,
                                shape=shape, transpose=transpose)
 
     assert (bm.allclose(r1, r2))
@@ -167,26 +169,26 @@ class Test_event_csr_matvec_taichi(parameterized.TestCase):
 
     # vmap 'data'
     events = bm.as_jax(rng.random(shape[0] if transpose else shape[1])) < 0.1
-    f1 = jax.vmap(partial(bm.event.csrmv, indices=indices, indptr=indptr, events=events,
+    f1 = jax.vmap(partial(brainpylib_csr_matvec, indices=indices, indptr=indptr, events=events,
                           shape=shape, transpose=transpose))
-    f2 = jax.vmap(partial(bm.event.csrmv_taichi, indices=indices, indptr=indptr, events=events,
+    f2 = jax.vmap(partial(taichi_csr_matvec, indices=indices, indptr=indptr, events=events,
                           shape=shape, transpose=transpose))
     vmap_data = bm.as_jax(rng.random((10, indices.shape[0])))
     self.assertTrue(bm.allclose(f1(vmap_data), f2(vmap_data)))
 
     # vmap 'events'
     data = bm.as_jax(rng.random(indices.shape))
-    f3 = jax.vmap(partial(bm.event.csrmv, data, indices, indptr,
+    f3 = jax.vmap(partial(brainpylib_csr_matvec, data, indices, indptr,
                           shape=shape, transpose=transpose))
-    f4 = jax.vmap(partial(bm.event.csrmv_taichi, data, indices, indptr,
+    f4 = jax.vmap(partial(taichi_csr_matvec, data, indices, indptr,
                           shape=shape, transpose=transpose))
     vmap_data = bm.as_jax(rng.random((10, shape[0] if transpose else shape[1]))) < 0.1
     self.assertTrue(bm.allclose(f3(vmap_data), f4(vmap_data)))
 
     # vmap 'data' and 'events'
-    f5 = jax.vmap(lambda dd, ee: bm.event.csrmv(dd, indices, indptr, ee,
+    f5 = jax.vmap(lambda dd, ee: brainpylib_csr_matvec(dd, indices, indptr, ee,
                                                 shape=shape, transpose=transpose))
-    f6 = jax.vmap(lambda dd, ee: bm.event.csrmv_taichi(dd, indices, indptr, ee,
+    f6 = jax.vmap(lambda dd, ee: taichi_csr_matvec(dd, indices, indptr, ee,
                                                        shape=shape, transpose=transpose))
     vmap_data1 = bm.as_jax(rng.random((10, indices.shape[0])))
     vmap_data2 = bm.as_jax(rng.random((10, shape[0] if transpose else shape[1]))) < 0.2
@@ -215,22 +217,22 @@ class Test_event_csr_matvec_taichi(parameterized.TestCase):
 
     # grad 'data'
     data = bm.as_jax(rng.random(indices.shape))
-    r1 = jax.grad(sum_op(bm.event.csrmv))(
+    r1 = jax.grad(sum_op(brainpylib_csr_matvec))(
       data, indices, indptr, events, shape=shape, transpose=transpose)
-    r2 = jax.grad(sum_op(bm.event.csrmv_taichi))(
+    r2 = jax.grad(sum_op(taichi_csr_matvec))(
       data, indices, indptr, events, shape=shape, transpose=transpose)
     self.assertTrue(bm.allclose(r1, r2))
 
     # grad 'events'
-    r3 = jax.grad(sum_op(bm.event.csrmv), argnums=3)(
+    r3 = jax.grad(sum_op(brainpylib_csr_matvec), argnums=3)(
       data, indices, indptr, events.astype(float), shape=shape, transpose=transpose)
-    r4 = jax.grad(sum_op(bm.event.csrmv_taichi), argnums=3)(
+    r4 = jax.grad(sum_op(taichi_csr_matvec), argnums=3)(
       data, indices, indptr, events.astype(float), shape=shape, transpose=transpose)
     self.assertTrue(bm.allclose(r3, r4))
 
-    r5 = jax.grad(sum_op(bm.event.csrmv), argnums=(0, 3))(
+    r5 = jax.grad(sum_op(brainpylib_csr_matvec), argnums=(0, 3))(
       data, indices, indptr, events.astype(float), shape=shape, transpose=transpose)
-    r6 = jax.grad(sum_op(bm.event.csrmv_taichi), argnums=(0, 3))(
+    r6 = jax.grad(sum_op(taichi_csr_matvec), argnums=(0, 3))(
       data, indices, indptr, events.astype(float), shape=shape, transpose=transpose)
     self.assertTrue(bm.allclose(r5[0], r6[0]))
     self.assertTrue(bm.allclose(r5[1], r6[1]))
