+++ brainpy/_src/math/jitconn/__init__.py
@@ -1,5 +1,3 @@
 
 from ._matvec import *
-from ._matvec_taichi import *
 from ._event_matvec import *
-from ._event_matvec_taichi import *
