+++ brainpy/_src/math/jitconn/tests/test_event_matvec.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+from functools import partial
 
 import jax
 import jax.numpy as jnp
@@ -19,6 +20,12 @@ shapes = [(100, 200),
           # (1000, 10),
           (1000, 2)]
 
+brainpylib_mv_prob_homo = partial(bm.jitconn.event_mv_prob_homo, method='brainpylib')
+taichi_mv_prob_homo = partial(bm.jitconn.event_mv_prob_homo, method='taichi')
+brainpylib_mv_prob_uniform = partial(bm.jitconn.event_mv_prob_uniform, method='brainpylib')
+taichi_mv_prob_uniform = partial(bm.jitconn.event_mv_prob_uniform, method='taichi')
+brainpylib_mv_prob_normal = partial(bm.jitconn.event_mv_prob_normal, method='brainpylib')
+taichi_mv_prob_normal = partial(bm.jitconn.event_mv_prob_normal, method='taichi')
 
 class Test_event_matvec_prob_conn(parameterized.TestCase):
   def __init__(self, *args, platform='cpu', **kwargs):
@@ -53,7 +60,7 @@ class Test_event_matvec_prob_conn(parameterized.TestCase):
     if not bool_event:
       events = events.astype(float)
 
-    r1 = bm.jitconn.event_mv_prob_homo(events,
+    r1 = brainpylib_mv_prob_homo(events,
                                        homo_data,
                                        conn_prob=prob,
                                        shape=shape,
@@ -62,7 +69,7 @@ class Test_event_matvec_prob_conn(parameterized.TestCase):
                                        transpose=transpose)
     r1 = jax.block_until_ready(r1)
 
-    r2 = bm.jitconn.event_mv_prob_homo(events,
+    r2 = brainpylib_mv_prob_homo(events,
                                        homo_data,
                                        conn_prob=prob,
                                        shape=shape,
@@ -72,7 +79,7 @@ class Test_event_matvec_prob_conn(parameterized.TestCase):
     r2 = jax.block_until_ready(r2)
     self.assertTrue(jnp.allclose(r1, r2))
 
-    r3 = bm.jitconn.event_mv_prob_homo(events,
+    r3 = brainpylib_mv_prob_homo(events,
                                        homo_data,
                                        conn_prob=prob,
                                        shape=(shape[1], shape[0]),
@@ -120,7 +127,7 @@ class Test_event_matvec_prob_conn(parameterized.TestCase):
     weights = bm.as_jax(rng.random(10))
 
     f1 = jax.vmap(
-      lambda event, data: bm.jitconn.event_mv_prob_homo(
+      lambda event, data: brainpylib_mv_prob_homo(
         event, data, conn_prob=prob, shape=shape, seed=seed,
         transpose=transpose, outdim_parallel=outdim_parallel
       )
@@ -164,7 +171,7 @@ class Test_event_matvec_prob_conn(parameterized.TestCase):
     events = events.astype(float)
 
     f1 = jax.grad(
-      lambda event, data: bm.jitconn.event_mv_prob_homo(
+      lambda event, data: brainpylib_mv_prob_homo(
         event, data, conn_prob=prob, shape=shape, seed=seed,
         outdim_parallel=outdim_parallel, transpose=transpose
       ).sum(),
@@ -231,7 +238,7 @@ class Test_event_matvec_prob_conn(parameterized.TestCase):
     if not bool_event:
       events = events.astype(float)
 
-    r1 = bm.jitconn.event_mv_prob_uniform(events,
+    r1 = brainpylib_mv_prob_uniform(events,
                                           w_low=w_low,
                                           w_high=w_high,
                                           conn_prob=prob,
@@ -241,7 +248,7 @@ class Test_event_matvec_prob_conn(parameterized.TestCase):
                                           transpose=transpose)
     r1 = jax.block_until_ready(r1)
 
-    r2 = bm.jitconn.event_mv_prob_uniform(events,
+    r2 = brainpylib_mv_prob_uniform(events,
                                           w_low=w_low,
                                           w_high=w_high,
                                           conn_prob=prob,
@@ -252,7 +259,7 @@ class Test_event_matvec_prob_conn(parameterized.TestCase):
     r2 = jax.block_until_ready(r2)
     self.assertTrue(jnp.allclose(r1, r2))
 
-    r3 = bm.jitconn.event_mv_prob_uniform(events,
+    r3 = brainpylib_mv_prob_uniform(events,
                                           w_low=w_low,
                                           w_high=w_high,
                                           conn_prob=prob,
@@ -302,7 +309,7 @@ class Test_event_matvec_prob_conn(parameterized.TestCase):
       events = events.astype(float)
 
     f1 = jax.vmap(
-      lambda e: bm.jitconn.event_mv_prob_uniform(e,
+      lambda e: brainpylib_mv_prob_uniform(e,
                                                  w_low=0.,
                                                  w_high=1.,
                                                  conn_prob=prob,
@@ -352,7 +359,7 @@ class Test_event_matvec_prob_conn(parameterized.TestCase):
     events = events.astype(float)
 
     f1 = jax.grad(
-      lambda e, w_high: bm.jitconn.event_mv_prob_uniform(
+      lambda e, w_high: brainpylib_mv_prob_uniform(
         e,
         w_low=0.,
         w_high=w_high,
@@ -413,7 +420,7 @@ class Test_event_matvec_prob_conn(parameterized.TestCase):
     if not bool_event:
       events = events.astype(float)
 
-    r1 = bm.jitconn.event_mv_prob_normal(events,
+    r1 = brainpylib_mv_prob_normal(events,
                                          w_mu=w_mu,
                                          w_sigma=w_sigma,
                                          conn_prob=prob,
@@ -423,7 +430,7 @@ class Test_event_matvec_prob_conn(parameterized.TestCase):
                                          transpose=transpose)
     r1 = jax.block_until_ready(r1)
 
-    r2 = bm.jitconn.event_mv_prob_normal(events,
+    r2 = brainpylib_mv_prob_normal(events,
                                          w_mu=w_mu,
                                          w_sigma=w_sigma,
                                          conn_prob=prob,
@@ -434,7 +441,7 @@ class Test_event_matvec_prob_conn(parameterized.TestCase):
     r2 = jax.block_until_ready(r2)
     self.assertTrue(jnp.allclose(r1, r2))
 
-    r3 = bm.jitconn.event_mv_prob_normal(events,
+    r3 = brainpylib_mv_prob_normal(events,
                                          w_mu=w_mu,
                                          w_sigma=w_sigma,
                                          conn_prob=prob,
@@ -486,7 +493,7 @@ class Test_event_matvec_prob_conn(parameterized.TestCase):
     if not bool_event:
       events = events.astype(float)
 
-    f1 = jax.vmap(lambda e: bm.jitconn.event_mv_prob_normal(e,
+    f1 = jax.vmap(lambda e: brainpylib_mv_prob_normal(e,
                                                             w_mu=0.,
                                                             w_sigma=1.,
                                                             conn_prob=prob,
@@ -536,7 +543,7 @@ class Test_event_matvec_prob_conn(parameterized.TestCase):
 
     f1 = jax.jit(
       jax.grad(
-        lambda e, w_sigma: bm.jitconn.event_mv_prob_normal(
+        lambda e, w_sigma: brainpylib_mv_prob_normal(
           e,
           w_mu=0.,
           w_sigma=w_sigma,
