+++ brainpy/_src/math/jitconn/tests/test_matvec_taichi.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-
+from functools import partial
 
 import jax
 import jax.numpy as jnp
@@ -10,6 +10,12 @@ import brainpy.math as bm
 shapes = [(100, 200), (10, 1000), (2, 1000), (1000, 10), (1000, 2)]
 shapes = [(100, 200), (2, 1000), (1000, 2)]
 
+brainpylib_mv_prob_homo = partial(bm.jitconn.mv_prob_homo, method='brainpylib')
+taichi_mv_prob_homo = partial(bm.jitconn.mv_prob_homo, method='taichi')
+brainpylib_mv_prob_uniform = partial(bm.jitconn.mv_prob_uniform, method='brainpylib')
+taichi_mv_prob_uniform = partial(bm.jitconn.mv_prob_uniform, method='taichi')
+brainpylib_mv_prob_normal = partial(bm.jitconn.mv_prob_normal, method='brainpylib')
+taichi_mv_prob_normal = partial(bm.jitconn.mv_prob_normal, method='taichi')
 
 class Test_matvec_prob_conn(parameterized.TestCase):
   def __init__(self, *args, platform='cpu', **kwargs):
@@ -51,7 +57,7 @@ class Test_matvec_prob_conn(parameterized.TestCase):
     rng = bm.random.RandomState()
     vector = bm.as_jax(rng.random(shape[0] if transpose else shape[1]))
 
-    r1 = bm.jitconn.mv_prob_homo_taichi(vector,
+    r1 = taichi_mv_prob_homo(vector,
                                         homo_data,
                                         conn_prob=prob,
                                         shape=shape,
@@ -59,7 +65,7 @@ class Test_matvec_prob_conn(parameterized.TestCase):
                                         outdim_parallel=outdim_parallel,
                                         transpose=transpose)
 
-    r2 = bm.jitconn.mv_prob_homo_taichi(vector,
+    r2 = taichi_mv_prob_homo(vector,
                                         homo_data,
                                         conn_prob=prob,
                                         shape=shape,
@@ -68,7 +74,7 @@ class Test_matvec_prob_conn(parameterized.TestCase):
                                         transpose=transpose)
     self.assertTrue(jnp.allclose(r1, r2))
 
-    r2 = bm.jitconn.mv_prob_homo_taichi(vector,
+    r2 = taichi_mv_prob_homo(vector,
                                         homo_data,
                                         conn_prob=prob,
                                         shape=(shape[1], shape[0]),
@@ -111,7 +117,7 @@ class Test_matvec_prob_conn(parameterized.TestCase):
     weights = bm.as_jax(rng.random(10))
 
     f1 = jax.vmap(
-      lambda event, data: bm.jitconn.mv_prob_homo_taichi(
+      lambda event, data: taichi_mv_prob_homo(
         event, data,
         conn_prob=prob, shape=shape, seed=seed,
         outdim_parallel=outdim_parallel, transpose=transpose
@@ -156,7 +162,7 @@ class Test_matvec_prob_conn(parameterized.TestCase):
     events = events.astype(float)
 
     f1 = jax.grad(
-      lambda event, data: bm.jitconn.mv_prob_homo_taichi(
+      lambda event, data: taichi_mv_prob_homo(
         event, data,
         conn_prob=prob,
         shape=shape,
@@ -213,7 +219,7 @@ class Test_matvec_prob_conn(parameterized.TestCase):
     rng = bm.random.RandomState()
     events = bm.as_jax(rng.random(shape[0] if transpose else shape[1]))
 
-    r1 = bm.jitconn.mv_prob_uniform_taichi(events,
+    r1 = taichi_mv_prob_uniform(events,
                                            w_low=w_low,
                                            w_high=w_high,
                                            conn_prob=prob,
@@ -222,7 +228,7 @@ class Test_matvec_prob_conn(parameterized.TestCase):
                                            outdim_parallel=outdim_parallel,
                                            transpose=transpose)
 
-    r2 = bm.jitconn.mv_prob_uniform_taichi(events,
+    r2 = taichi_mv_prob_uniform(events,
                                            w_low=w_low,
                                            w_high=w_high,
                                            conn_prob=prob,
@@ -235,7 +241,7 @@ class Test_matvec_prob_conn(parameterized.TestCase):
       print(r1, r2)
     self.assertTrue(c)
 
-    r2 = bm.jitconn.mv_prob_uniform_taichi(events,
+    r2 = taichi_mv_prob_uniform(events,
                                            w_low=w_low,
                                            w_high=w_high,
                                            conn_prob=prob,
@@ -281,7 +287,7 @@ class Test_matvec_prob_conn(parameterized.TestCase):
     rng = bm.random.RandomState()
     events = bm.as_jax(rng.random((10, shape[0] if transpose else shape[1])))
 
-    f1 = jax.vmap(lambda e: bm.jitconn.mv_prob_uniform_taichi(e,
+    f1 = jax.vmap(lambda e: taichi_mv_prob_uniform(e,
                                                               w_low=0.,
                                                               w_high=1.,
                                                               conn_prob=prob,
@@ -330,7 +336,7 @@ class Test_matvec_prob_conn(parameterized.TestCase):
     events = bm.as_jax(rng.random(shape[0] if transpose else shape[1]))
 
     f1 = jax.grad(
-      lambda e, w_low, w_high: bm.jitconn.mv_prob_uniform_taichi(
+      lambda e, w_low, w_high: taichi_mv_prob_uniform(
         e,
         w_low=w_low,
         w_high=w_high,
@@ -390,7 +396,7 @@ class Test_matvec_prob_conn(parameterized.TestCase):
     rng = bm.random.RandomState()
     events = bm.as_jax(rng.random(shape[0] if transpose else shape[1]))
 
-    r1 = bm.jitconn.mv_prob_normal_taichi(events,
+    r1 = taichi_mv_prob_normal(events,
                                           w_mu=w_mu,
                                           w_sigma=w_sigma,
                                           conn_prob=prob,
@@ -399,7 +405,7 @@ class Test_matvec_prob_conn(parameterized.TestCase):
                                           outdim_parallel=outdim_parallel,
                                           transpose=transpose)
 
-    r2 = bm.jitconn.mv_prob_normal_taichi(events,
+    r2 = taichi_mv_prob_normal(events,
                                           w_mu=w_mu,
                                           w_sigma=w_sigma,
                                           conn_prob=prob,
@@ -412,7 +418,7 @@ class Test_matvec_prob_conn(parameterized.TestCase):
       print(r1, r2)
     self.assertTrue(c)
 
-    r2 = bm.jitconn.mv_prob_normal_taichi(events,
+    r2 = taichi_mv_prob_normal(events,
                                           w_mu=w_mu,
                                           w_sigma=w_sigma,
                                           conn_prob=prob,
@@ -459,7 +465,7 @@ class Test_matvec_prob_conn(parameterized.TestCase):
     rng = bm.random.RandomState()
     events = bm.as_jax(rng.random((10, shape[0] if transpose else shape[1])))
 
-    f1 = jax.vmap(lambda e: bm.jitconn.mv_prob_normal_taichi(e,
+    f1 = jax.vmap(lambda e: taichi_mv_prob_normal(e,
                                                              w_mu=0.,
                                                              w_sigma=1.,
                                                              conn_prob=prob,
@@ -512,7 +518,7 @@ class Test_matvec_prob_conn(parameterized.TestCase):
     events = events.astype(float)
 
     f1 = jax.grad(
-      lambda e, w_sigma: bm.jitconn.mv_prob_normal_taichi(
+      lambda e, w_sigma: taichi_mv_prob_normal(
         e,
         w_mu=0.,
         w_sigma=w_sigma,
