+++ brainpy/_src/math/sparse/__init__.py
@@ -1,7 +1,6 @@
 
 from ._coo_mv import *
 from ._csr_mv import *
-from ._csr_mv_taichi import *
 from ._utils import *
 from ._bsr_mv import *
 from ._bsr_mm import *
