+++ brainpy/math/event.py
@@ -1,6 +1,5 @@
 
 from brainpy._src.math.event import (
   csrmv as csrmv,
-  csrmv_taichi as csrmv_taichi,
   info as info,
 )
