+++ brainpy/math/jitconn.py
@@ -6,13 +6,5 @@ from brainpy._src.math.jitconn import (
   mv_prob_homo as mv_prob_homo,
   mv_prob_uniform as mv_prob_uniform,
   mv_prob_normal as mv_prob_normal,
-
-  event_mv_prob_homo_taichi as event_mv_prob_homo_taichi,
-  event_mv_prob_uniform_taichi as event_mv_prob_uniform_taichi,
-  event_mv_prob_normal_taichi as event_mv_prob_normal_taichi,
-
-  mv_prob_homo_taichi as mv_prob_homo_taichi,
-  mv_prob_uniform_taichi as mv_prob_uniform_taichi,
-  mv_prob_normal_taichi as mv_prob_normal_taichi
 )
